# Trial to speed up mixin column info.