@taldcroft - just a proof of concept, to try to make getting `col.info` itself faster. More generally, we might want to think about replacing getting/setting attributes on the parent not via `__getattr__` and `__setattr__`, but rather via (automatically generated) properties on `info`.

```
from astropy.table import MaskedColumn
m = MaskedColumn([1, 2])
%timeit m.info
# 1000000 loops, best of 5: 336 ns per loop
# on current master: 1.26 µs
m.info._name = 'bla'
%timeit m.info._name
# 1000000 loops, best of 5: 370 ns per loop
# on current master: 1.28 µs
%timeit m.name
# 10000000 loops, best of 5: 82.3 ns per loop
# "only" 4.5 times faster instead of 15!
```