@taldcroft 2019-07-09
So this inspired me to start thinking outside our current box.  The very first thing about `DataInfo` is that it is a descriptor class that is storing the per-instance attribute information on the parent object and always dynamically finding the parent.  You made that faster, which is definitely good.  Overall `DataInfo` is a bit complicated and has all these redirects and knobs.

Is there any reasonable way we can scrap that whole paradigm and simplify a whole lot?  Just make `info` a property pointing to an ordinary object stored in `_info`.  The info "attributes" are properties that allow quite fast access, with setters that do validation as necessary.

Maybe we've been dancing around this idea, but I wonder if something like below could fly?  It might end up making the `DataInfo` concept much simpler, cleaner, and faster.

https://github.com/astropy/astropy/compare/master...taldcroft:table-info-play
```
In [2]: c = Column([1,2], name='hello')

In [3]: timeit c.name
165 ns ± 0.627 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)

In [4]: timeit c.finfo.name
305 ns ± 2.74 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [5]: timeit c.info.name
3.89 µs ± 56.5 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```