@mhvk 2019-07-09
@taldcroft - interesting idea. Part of this is what I had in mind by auto-generation - one can just generate the `name` properties, etc., by a metaclass from `attrs_from_parent`, etc. I'm not sure making `info` a property is going to gain much over it being a descriptor (a property is a descriptor, after all), but I think making it simpler/more permanent is good, and I like especially the much closer link to the `parent` that you have (though obviously we need to look into why we had that `weakref` in the first place! I experimented with not setting `_parent` on every call but got quite a few errors).

p.s. The good part is that we've got lots of tests, so one can quickly find out whether a new idea will work!