@mhvk 2019-07-15
closing in favour of #8998 (which includes this commit, though it partially undoes it)